Moved cache related functionalities to separate package
Also defined staging there
